create exel-data-viewer component
